requirements.txt
aiosqlite==0.21.0
anyio==4.9.0
black==25.1.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
mypy_extensions==1.1.0
nest-asyncio==1.6.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
python-dotenv==1.1.1
python-telegram-bot==22.2
requests==2.32.4
sniffio==1.3.1
SQLAlchemy==2.0.41
typing_extensions==4.14.0
urllib3==2.5.0
